Bug fix when migrating relationships that have been renamed
The old code would unconditionally look at the new name being used, and would attempt to locate the relationship in the source entity by the new name. This would only work if the relationship's name has not been changed. As per recommendation by Apple's document, the user is expected to use Renaming ID found in the Data Model when renaming attributes/relationships. The code fix will first attempt to read from the renamingIdentifier property in the destination source. If no such information can be found, we can make the assumption that the relationship's name has not changed and will use it as is.